chatbot_backend/requirements.txt
# Django and REST Framework
Django==5.2
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Machine Learning and NLP
tensorflow==2.18.0
nltk==3.9.1
spacy==3.8.2
numpy==1.26.4
scikit-learn==1.5.2

# Data Processing
pandas==2.2.3
pickle-mixin==1.0.2

# HTTP Requests
requests==2.32.3

# Utilities
python-dateutil==2.9.0.post0
pathlib2==2.3.7.post1

# Development
python-dotenv==1.0.1